requirements.txt
alembic==1.6.5
aniso8601==7.0.0
anyio==3.6.1
asgiref==3.5.1
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.0
bokeh==2.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
coloredlogs==14.0
croniter==1.3.5
dagit==0.14.15
dagster==0.14.15
dagster-graphql==0.14.15
dask==2022.5.0
dask-glm==0.2.0
dask-ml==2022.1.22
defusedxml==0.7.1
distributed==2022.5.0
docstring-parser==0.14.1
entrypoints==0.4
fastjsonschema==2.15.3
fsspec==2022.3.0
gql==2.0.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
grpcio==1.46.1
grpcio-health-checking==1.43.0
h11==0.13.0
HeapDict==1.0.1
httptools==0.4.0
humanfriendly==10.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
llvmlite==0.38.0
locket==1.0.0
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.0.1
mistune==0.8.4
msgpack==1.0.3
multipledispatch==0.6.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pandas-schema==0.3.6
pandocfilters==1.5.0
partd==1.2.0
pendulum==2.1.2
pep562==1.1
Pillow==9.1.0
pip==21.3.1
promise==2.3
protobuf==3.20.1
psutil==5.9.0
Pygments==2.12.0
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
pytzdata==2020.1
pywin32==304
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
Rx==1.6.1
scikit-learn==1.1.0
scipy==1.8.0
setuptools==60.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
starlette==0.20.0
tabulate==0.8.9
tblib==1.7.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.11.2
toposort==1.7
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typing-compat==0.1.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchdog==2.1.7
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
wheel==0.37.1
zict==2.2.0
